Stabilize 0.8.1 release gates

+++ apps/pi-zero/src/user_data_transfer_http.rs
@@ -392,7 +392,7 @@ fn available_space(path: &Path) -> Option<u64> {
         return None;
     }
     let stats = unsafe { stats.assume_init() };
-    Some((stats.f_bavail as u64).saturating_mul(stats.f_frsize as u64))
+    Some(stats.f_bavail.saturating_mul(stats.f_frsize))
 }
 
 #[cfg(not(unix))]
